+++ cmd/crd.go
@@ -37,6 +37,10 @@ import (
 	"github.com/arangodb/kube-arangodb/pkg/util/shutdown"
 )
 
+const (
+	AllSchemasValue = "all"
+)
+
 var (
 	cmdCRD = &cobra.Command{
 		Use:   "crd",
@@ -48,12 +52,18 @@ var (
 		Run:   cmdCRDInstallRun,
 		Short: "Install and update all required CRDs",
 	}
+	cmdCRDGenerate = &cobra.Command{
+		Use:   "generate",
+		Run:   cmdCRDGenerateRun,
+		Short: "Generates YAML of all required CRDs",
+	}
 )
 
 var (
 	crdInstallOptions struct {
-		validationSchema []string
-		force            bool
+		validationSchema      []string
+		preserveUnknownFields []string
+		force                 bool
 	}
 )
 
@@ -65,39 +75,107 @@ func init() {
 	cmdMain.AddCommand(cmdCRD)
 	cmdOps.AddCommand(cmdCRD)
 
-	f := cmdCRDInstall.Flags()
+	f := cmdCRD.PersistentFlags()
 	f.StringArrayVar(&crdInstallOptions.validationSchema, "crd.validation-schema", defaultValidationSchemaEnabled, "Controls which CRD should have validation schema <crd-name>=<true/false>.")
+	f.StringArrayVar(&crdInstallOptions.preserveUnknownFields, "crd.preserve-unknown-fields", nil, "Controls which CRD should have enabled preserve unknown fields in validation schema <crd-name>=<true/false>.")
 	f.BoolVar(&crdInstallOptions.force, "crd.force-update", false, "Enforce CRD Schema update")
+
 	cmdCRD.AddCommand(cmdCRDInstall)
+	cmdCRD.AddCommand(cmdCRDGenerate)
 }
 
-func prepareCRDOptions(schemaEnabledArgs []string) (map[string]crds.CRDOptions, error) {
+func prepareCRDOptions(schemaEnabledArgs []string, preserveUnknownFieldsArgs []string) (map[string]crds.CRDOptions, error) {
 	defaultOptions := crd.GetDefaultCRDOptions()
-	result := make(map[string]crds.CRDOptions)
 	var err error
-	for _, arg := range schemaEnabledArgs {
-		parts := strings.Split(arg, "=")
 
-		crdName := parts[0]
-		opts, ok := defaultOptions[crdName]
-		if !ok {
-			return nil, fmt.Errorf("unknown CRD %s", crdName)
-		}
+	schemaEnabled := map[string]bool{}
+	preserveUnknownFields := map[string]bool{}
+
+	for _, arg := range schemaEnabledArgs {
+		parts := strings.SplitN(arg, "=", 2)
+
+		var enabled bool
 
 		if len(parts) == 2 {
-			opts.WithSchema, err = strconv.ParseBool(parts[1])
+			enabled, err = strconv.ParseBool(parts[1])
 			if err != nil {
 				return nil, errors.Wrapf(err, "not a bool value: %s", parts[1])
 			}
+
 		}
 
-		result[crdName] = opts
+		schemaEnabled[parts[0]] = enabled
 	}
-	return result, nil
+
+	for _, arg := range preserveUnknownFieldsArgs {
+		parts := strings.SplitN(arg, "=", 2)
+
+		var enabled bool
+
+		if len(parts) == 2 {
+			enabled, err = strconv.ParseBool(parts[1])
+			if err != nil {
+				return nil, errors.Wrapf(err, "not a bool value: %s", parts[1])
+			}
+
+		}
+
+		preserveUnknownFields[parts[0]] = enabled
+	}
+
+	for k := range schemaEnabled {
+		if k == AllSchemasValue {
+			continue
+		}
+		if _, ok := defaultOptions[k]; !ok {
+			return nil, fmt.Errorf("unknown CRD %s", k)
+		}
+	}
+
+	for k := range preserveUnknownFields {
+		if k == AllSchemasValue {
+			continue
+		}
+		if _, ok := defaultOptions[k]; !ok {
+			return nil, fmt.Errorf("unknown CRD %s", k)
+		}
+	}
+
+	// Override the defaults
+	if v, ok := schemaEnabled[AllSchemasValue]; ok {
+		delete(preserveUnknownFields, AllSchemasValue)
+		for k := range defaultOptions {
+			z := defaultOptions[k]
+			z.WithSchema = v
+			defaultOptions[k] = z
+		}
+	}
+	if v, ok := preserveUnknownFields[AllSchemasValue]; ok {
+		delete(preserveUnknownFields, AllSchemasValue)
+		for k := range defaultOptions {
+			z := defaultOptions[k]
+			z.WithPreserve = v
+			defaultOptions[k] = z
+		}
+	}
+
+	// Set explicit words
+	for k, v := range schemaEnabled {
+		z := defaultOptions[k]
+		z.WithSchema = v
+		defaultOptions[k] = z
+	}
+	for k, v := range preserveUnknownFields {
+		z := defaultOptions[k]
+		z.WithPreserve = v
+		defaultOptions[k] = z
+	}
+
+	return defaultOptions, nil
 }
 
 func cmdCRDInstallRun(cmd *cobra.Command, args []string) {
-	crdOpts, err := prepareCRDOptions(crdInstallOptions.validationSchema)
+	crdOpts, err := prepareCRDOptions(crdInstallOptions.validationSchema, crdInstallOptions.preserveUnknownFields)
 	if err != nil {
 		logger.Fatal("Invalid --crd.validation-schema args: %s", err)
 		return
@@ -113,6 +191,19 @@ func cmdCRDInstallRun(cmd *cobra.Command, args []string) {
 	defer cancel()
 
 	err = crd.EnsureCRDWithOptions(ctx, client, crd.EnsureCRDOptions{IgnoreErrors: false, CRDOptions: crdOpts, ForceUpdate: crdInstallOptions.force})
+	if err != nil {
+		os.Exit(1)
+	}
+}
+
+func cmdCRDGenerateRun(cmd *cobra.Command, args []string) {
+	crdOpts, err := prepareCRDOptions(crdInstallOptions.validationSchema, crdInstallOptions.preserveUnknownFields)
+	if err != nil {
+		logger.Fatal("Invalid --crd.validation-schema args: %s", err)
+		return
+	}
+
+	err = crd.GenerateCRDYAMLWithOptions(crd.EnsureCRDOptions{IgnoreErrors: false, CRDOptions: crdOpts, ForceUpdate: crdInstallOptions.force}, cmd.OutOrStdout())
 	if err != nil {
 		os.Exit(1)
 	}