+++ pkg/apis/deployment/v1/server_group.go
@@ -1,7 +1,7 @@
 //
 // DISCLAIMER
 //
-// Copyright 2016-2022 ArangoDB GmbH, Cologne, Germany
+// Copyright 2016-2024 ArangoDB GmbH, Cologne, Germany
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -38,6 +38,12 @@ func (s ServerGroups) Contains(group ServerGroup) bool {
 }
 
 type ServerGroup int
+
+func (g *ServerGroup) OpenAPISchemaType() []string {
+	return []string{
+		"string",
+	}
+}
 
 func (g *ServerGroup) UnmarshalJSON(bytes []byte) error {
 	if bytes == nil {